Index action will only show items for that user
Creates item-index.sh curl script for testing.
Made changes to the index action in the item controller. Now passing
the req.user.id to the .Find method -- to link the index action to only
the user who is signed in. Adds the authenticate before action
for index. Index, create, update, and destroy have all been tested with
curl.